Added Access Restrictions to Vehicle Controller

+++ src/main/java/truckmanagementproject/web/controllers/vehicles/VehicleController.java
@@ -6,11 +6,14 @@ import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
+import truckmanagementproject.services.services.auth.AuthService;
 import truckmanagementproject.services.services.vehicles.VehicleService;
 import truckmanagementproject.services.models.vehicles.AddVehicleServiceModel;
+import truckmanagementproject.web.models.auth.LoginUserViewModel;
 import truckmanagementproject.web.models.vehicles.AddVehicleModel;
 import truckmanagementproject.web.models.vehicles.VehicleViewModel;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -26,16 +29,21 @@ public class VehicleController {
 
     private final ModelMapper mapper;
     private final VehicleService vehicleService;
+    private final AuthService authService;
 
     @Autowired
-    public VehicleController(ModelMapper mapper, VehicleService vehicleService) {
+    public VehicleController(ModelMapper mapper, VehicleService vehicleService, AuthService authService) {
         this.mapper = mapper;
         this.vehicleService = vehicleService;
+        this.authService = authService;
     }
 
-    //TODO -> ONLY ACCESSIBLE FROM ADMIN
     @GetMapping("/add")
-    public String getAddVehicleForm(@ModelAttribute("model") AddVehicleModel model) {
+    public String getAddVehicleForm(@ModelAttribute("model") AddVehicleModel model, HttpSession session) throws Exception {
+        LoginUserViewModel user = (LoginUserViewModel) session.getAttribute("user");
+        if (!authService.isUserAdmin(user)) {
+            throw new Exception("Unauthorized user");
+        }
         return "vehicles/add-vehicle";
     }
 
@@ -55,9 +63,12 @@ public class VehicleController {
         }
     }
 
-    //TODO -> ONLY ACCESSIBLE FROM ADMIN + MANAGER
     @GetMapping("/all")
-    public ModelAndView getAllVehicles(ModelAndView modelAndView) {
+    public ModelAndView getAllVehicles(ModelAndView modelAndView, HttpSession session) throws Exception {
+        LoginUserViewModel user = (LoginUserViewModel) session.getAttribute("user");
+        if (!authService.isUserAdmin(user) && !authService.isUserManager(user)) {
+            throw new Exception("Unauthorized user");
+        }
         List<VehicleViewModel> vehicles = vehicleService.getAllVehicles()
                 .stream()
                 .map(vehicle -> mapper.map(vehicle, VehicleViewModel.class))
@@ -68,9 +79,12 @@ public class VehicleController {
         return modelAndView;
     }
 
-    //TODO -> ONLY ACCESSIBLE FROM ADMIN
     @GetMapping("/remove/{id}")
-    public ModelAndView removeDriverDoc(@PathVariable String id) {
+    public ModelAndView removeDriverDoc(@PathVariable String id, HttpSession session) throws Exception {
+        LoginUserViewModel user = (LoginUserViewModel) session.getAttribute("user");
+        if (!authService.isUserAdmin(user)) {
+            throw new Exception("Unauthorized user");
+        }
         vehicleService.removeVehicle(id);
         return new ModelAndView("redirect:/vehicles/all");
     }
